+++ popkorn/src/iosMain/kotlin/cc/popkorn/Platform.kt
@@ -18,10 +18,10 @@ import kotlin.reflect.KClass
 
 
 /**
- * Implementation for Native of the methods/classes that are Platform-dependent
+ * Implementation for Ios of the methods/classes that are Platform-dependent
  *
  * @author Pau Corbella
- * @since 1.6.0
+ * @since 2.0.0
  */
 
 actual typealias WeakReference<T> = kotlin.native.ref.WeakReference<T>