+++ popkorn/src/jsMain/kotlin/cc/popkorn/Platform.kt
@@ -13,7 +13,7 @@ import kotlin.reflect.KClass
  * Implementation for JS of the methods/classes that are Platform-dependent
  *
  * @author Pau Corbella
- * @since 1.6.0
+ * @since 2.0.0
  */