+++ popkorn/src/jsMain/kotlin/cc/popkorn/PopKornCompat.kt
@@ -7,7 +7,7 @@ import cc.popkorn.mapping.Mapping
  * Compatibility class to use PopKorn from js code
  *
  * @author Pau Corbella
- * @since 1.6.0
+ * @since 2.0.0
  */