+++ popkorn/src/jvmMain/kotlin/cc/popkorn/Platform.kt
@@ -10,7 +10,7 @@ import kotlin.reflect.KClass
  * Implementation for JVM of the methods/classes that are Platform-dependent
  *
  * @author Pau Corbella
- * @since 1.6.0
+ * @since 2.0.0
  */
 
 actual typealias WeakReference<T> = java.lang.ref.WeakReference<T>