+++ popkorn/src/nativeMain/kotlin/cc/popkorn/Platform.kt
@@ -14,7 +14,7 @@ import kotlin.reflect.KClass
  * Implementation for Native of the methods/classes that are Platform-dependent
  *
  * @author Pau Corbella
- * @since 1.6.0
+ * @since 2.0.0
  */