+++ popkorn/src/nativeMain/kotlin/cc/popkorn/PopKornCompat.kt
@@ -4,10 +4,10 @@ import cc.popkorn.mapping.Mapping
 
 
 /**
- * Compatibility class to use PopKorn from js code
+ * Compatibility class to use PopKorn from native code
  *
  * @author Pau Corbella
- * @since 1.6.0
+ * @since 2.0.0
  */